境界層業務概念改修

+++ DB/DBA/standard/service/src/main/java/jp/co/ndensan/reams/db/dba/service/core/kyokaisogaitosha/kyokaisohokenryodankai/KyokaisoHokenryoDankaiManager.java
@@ -9,9 +9,9 @@ import java.util.ArrayList;
 import java.util.List;
 import static java.util.Objects.requireNonNull;
 import jp.co.ndensan.reams.db.dba.business.core.kyokaisogaitosha.kyokaisohokenryodankai.KyokaisoHokenryoDankai;
-import jp.co.ndensan.reams.db.dba.entity.db.basic.DbT1007KyokaisoHokenryoDankaiEntity;
-import jp.co.ndensan.reams.db.dba.persistence.db.basic.DbT1007KyokaisoHokenryoDankaiDac;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.HihokenshaNo;
+import jp.co.ndensan.reams.db.dbz.entity.db.basic.DbT1007KyokaisoHokenryoDankaiEntity;
+import jp.co.ndensan.reams.db.dbz.persistence.db.basic.DbT1007KyokaisoHokenryoDankaiDac;
 import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYearMonth;
 import jp.co.ndensan.reams.uz.uza.math.Decimal;
@@ -65,8 +65,8 @@ public class KyokaisoHokenryoDankaiManager {
 
         DbT1007KyokaisoHokenryoDankaiEntity entity = dac.selectByKey(
                 被保険者番号,
-                履歴番号,
-                リンク番号,
+                履歴番号.intValue(),
+                リンク番号.intValue(),
                 適用開始年月);
         if (entity == null) {
             return null;
